requirements.txt
Pillow==10.4.0
imagehash==4.3.1
opencv-python==4.10.0.84
customtkinter==5.2.2
numpy
